refactor(render): split the terrain chunk generator's compute half out of terrain.ts
Step 1 of moving terrain generation off the main thread. Pure move, no
behaviour change: tests/terrain_chunk_geometry.test.ts hashes position, normal,
color, uv and index across all 36 Eastbrook chunks and is unchanged, and the
moved text is byte-identical to what left terrain.ts apart from four added
export keywords.

The generator was already shaped for this. beginChunkGeometry and the two row
fills write plain Float32Array / Uint32Array and touch no WebGL; only
finishChunkGeometry needs Three's BufferGeometry, and it is twelve lines. So
the seam falls between them, and the whole compute half plus its palette moves
out: lines 145 to 773 were already contiguous and self-contained, with just
SKIRT_DROP and SLOPE_EPS to bring along and only four symbols crossing the
boundary.

terrain.ts drops from 1621 lines to 992.

DELIBERATELY NOT NAMED *_core, and it still imports three. The vertex sampler
does its colour work with THREE.Color over a twenty-odd entry palette, and
THREE.Color applies an sRGB to linear conversion under ColorManagement.
Swapping that for plain float maths would reshade the entire world while
changing nothing's shape. That belongs in its own step, guarded by the geometry
pin, and until then a *_core name would be swept into RENDER_PURE_CORES and
fail the Three-free purity scan for a reason that has nothing to do with
whether the split is right. Three's maths classes run fine in a worker; they
only cost bundle size.

Why this is worth doing at all: generating a zone's geometry is 2 to 4 seconds
of arithmetic, but the background lane stretches it to 60 to 105 to avoid
stealing frame time. Frostveil measured terrainMs 3914 on the gating path and
85523 on the idle one for the same geometry, a 22x pacing tax. Off-thread it
disappears, and the outdoor fog clamp stops waiting on it.
